run: no more redirect stdout to /dev/null
Growing log files in container itself might an issue with many
registration, queries and high verbosity level.
Let's the user the option to keep stdout of main process, run the
container in background and `docker logs`for monitoring. In this way,
the log resources can be accessed and managed from the host, which is
good in a large setup, e.g.  logstash, docker-compose, multiple
registry.