Use github secret for image name.
Using a github secret for the image name is a bit easier to maintain for forks of the repo and fits in with use of secrets for other items in the CI build.

Also added a workflow_dispatch so the build can be started manually.  Useful for situations when debugging the actual CI/CD workflow.